refactor(nuqs): Rename files and remove ParserBuilder type annotations
Simplifies type signatures by relying on type inference from createParser.
Renames files to match the primary export name (withStorage).

Changes:
- Removed ParserBuilder<T> import and type annotations
- Use simple inline parser type: { parse, serialize }
- Rely on TypeScript inference from createParser<T>
- Renamed withLocalStorage.tsx → withStorage.tsx
- Renamed withLocalStorage.spec.tsx → withStorage.spec.tsx
- Renamed withLocalStorage.example.tsx → withStorage.example.tsx
- Updated all imports and references in README

Benefits:
- Simpler, cleaner type signatures
- Better file naming (matches primary export)
- Less verbose without losing type safety
- TypeScript infers return type automatically

All tests passing ✅ (35 tests)
No type errors ✅
